Remove(expenses): api, hook, interface 파일 분리, 경로 수정
