+++ src/hooks/useReportQuery.tsx
@@ -1,20 +1,9 @@
 "use client";
 
+import { getRevenueData } from "@/lib/reportApi";
 import { useQuery } from "@tanstack/react-query";
-import { getExpensesData, getExpensesDetailData, getRevenueData } from "@/lib/reportApi";
-import { exponseDetailDTOType, exponseDTOType } from "@/interface/report";
 
 export const useRevenueQuery = () => useQuery({
   queryKey: ["revenueData"],
   queryFn: getRevenueData,
-})
-
-export const useExpensesData = () => useQuery<exponseDTOType>({
-  queryKey: ["expensesData"],
-  queryFn: getExpensesData,
-})
-
-export const useExpensesDetailData = () => useQuery<exponseDetailDTOType>({
-  queryKey: ["expensesDetailData"],
-  queryFn: getExpensesDetailData,
 })
